pyproject.toml
[project]
name = "franchise-billing-platform"
version = "0.1.0"
description = "Cloud-first franchise billing backend for car detailing showrooms."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi>=0.116.0",
    "pydantic-settings>=2.6.0",
    "pydantic[email]>=2.10.0",
    "psycopg[binary]>=3.2.0",
    "pyjwt>=2.10.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.37",
    "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
